loading quadrics to objects

+++ Code/PathTracerSol/PathTracer/Material.h
@@ -15,7 +15,7 @@ public:
 	int  n;// Specular reflection exponent
 
 	Material(const double &r, const double &b, const double &g,
-		const double &Ka, const double &Kd, const double &Ks, const double &Kt, const double &n):
+		const double &Ka, const double &Kd, const double &Ks, const double &Kt, const int &n):
 	r(r), b(b), g(g), Ka(Ka), Kd(Kd), Ks(Ks), Kt(Kt), n(n) {}
 	Material(){}
 	~Material(){}
